Starts allocating pages for component data
Committed this WIP because I need to change direction. I need runtime
access to the archetype for a page, but this design requires casting
to it which means it needs to be known comptime.

Instead, all pages should be the same time, and just have some buffer
in which I can allocate the component data regardless of type.

They'll also store the current archetype and dynamically do the right
thing given the archetype.