Starts refactoring internal memory management api boundary
I want to rip out the internals of how memory is managed in the
ecs and replace it with something simpler. This is a big change,
though, so I'm first making the boundary between memory management
and the rest of the ecs logic cleaner.

I don't care if under the barrier things are messy, or if these
changes hurt performance, they're just setting the stage to make
the later refactor easy to do where all the details will be replaced
anyway.