Separates input system
Now that we can instantiate prefabs, if we wanted NPC controlled
ships all we'd need to do is replace player index with a union that
is either a player index OR a handle to an entity that stores that
NPC's input state. This would be a handle so that ships made of
multiple entities can all reference the same input state.