Starts moving away from trig, gets features like the double turret on kevin working again
WRT math: favoring vector math over trigonometry where possible. This tends to be easier to
work with. Most transformations will get simpler, a few won't but there are good abstractions
for them. Midway through right now so not a great demo of why this is worthwhile yet, but I've
done this both ways in the past and the non trig based approach always ends up worthwhile.