Moves post processing to compute
We're basically doing the same math right now since it's basically
a direct port of the old shader, so there doesn't appear to be
much advantage. However, now that it's ported, it opens the door to
better techniques. e.g. there are some really cool blur strategies
that aren't very workable in the graphics pipeline.