Uses bindless in compute, first pass at a Kawase blur, combines vert/frag stages
Bindless seems like overkill in compute since there are so many fewer
passes, but it's not about perf. Managing descriptor sets is a nightmare
not because it's complex, or even slow in this case, but because adding
a new effect requires modifying too many disparate pieces of code.

To add a multi pass blur you'd need to define a descriptor set layout for
it, create the actual sets per frame in flight or out of a pool you reset
per frame in flight, manage all the writes, etc. Huge overhead when you're
just trying to be creative and try out ideas.

OTOH, with bindless, you can just use one descriptor set for your entire
app--compute and graphics can share it--and you just use a few push constants
to select your inputs/outputs and control magnitudes, if you need more data
you use the push constants to index a buffer with more info.

Night and day. You can write fast, low level code, but still experiment without
drowning in boilerplate.

For similar reasons, vert and frag shaders are now combined. It's not conducive
to iteration to have these stages that are obviously going to be tightly coupled
in separate files, I don't know why GLSL prefers that.