Adds a moving average blur
Just experimenting at this point. Moving average blurs are good for
very large blur kernels because they are linear per pixel + a fixed
cost per block, allowing for very very large blur kernels that would
otherwise be impossible.

These are box blurs, but you can iterate 2 or 3 times to get a very
good guassian approximation. This is also much more controllable than
something like a Kawase blur which you need to empirically figure out
the iterations for to match a gaussian, whcih is very useful for making
the blur radius resolution independent. However, it probably needs to
be profiled a little more carefully on varied hardware to be shippable,
and one could always downsample to a fixed resolution before applying
Kawase to make it controllable.

This blur could be optimized further by using the sampler trick to read
two samples at a time. The improvements would be marginal since it would
only affect the fixed part of the cost.