Fixes reading uninitialized memory
Often these values were 0 which hid the issue, but sometimes they wouldn't
be and it lead to an intermittent bug. Easy catch with RenderDoc, integration
was worth it--should stop using preamble and baked in version since it doesn't
work well with RenderDoc.

Can experiment with just using #line instead for this?